+++ packages/dso-toolkit/src/components/icon-button/icon-button.models.ts
@@ -13,4 +13,5 @@ export interface IconButton {
   tooltipPlacement?: IconButtonTooltipPlacement;
   disabled?: boolean;
   dsoClick?: HandlerFunction;
+  toggled?: boolean;
 }